カラムの型修正

+++ db/migrate/20210318060334_create_weights.rb
@@ -4,7 +4,7 @@ class CreateWeights < ActiveRecord::Migration[6.0]
 
       t.timestamps
       t.references :user,               null: false, foreign_key: true
-      t.integer    :weight,             null: false
+      t.float      :weight,             null: false
       t.date       :date,        null: false
     end
   end
